cloudformation/docker/mosaic_meddra_coding/api/requirements.txt
Flask==1.0.2
numpy==1.16.3
torch==1.1.0
pytorch-nlp==0.4.1
requests==2.21.0
joblib==0.13.2
#regex==2019.4.14
gunicorn==19.9.0
gevent
